requirements.txt
PyGithub==2.6.1
pyadaptivecards==0.1.1
